requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
bcrypt==4.1.1
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
